requirements.txt
# Core dependencies for Claude API
anthropic>=0.34.0

# Bot Framework dependencies
botbuilder-core>=4.14.0
botbuilder-schema>=4.14.0
botframework-connector>=4.14.0
botbuilder-azure>=4.14.0
botbuilder-integration-aiohttp>=4.14.0

# CLI dependencies (for standalone usage)
rich>=13.0.0
click>=8.0.0
pyperclip>=1.8.0  # Optional: for clipboard support

# Async HTTP client
aiohttp>=3.8.0

# Utilities
python-dotenv>=0.19.0
requests>=2.25.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-mock>=3.10.0

# Development dependencies
black>=23.0.0
pylint>=2.15.0